Implement checkDomainRequest API call
This also rewrites the convenience methods check and check_many to be
based on the new check_domain_request function.

The reason expose the raw request method as well is premium domains.
The result can have an extra attribute that cannot be exposed through
the convenience methods while maintaining the API.